requirements-dev.txt
chardet
atlassian-python-api
pandas
PyYAML
datetime
pytz
flake8>=6.0.0
flake8-docstrings>=1.7.0
pydocstyle>=6.3.0
pre-commit>=3.0.0
sphinx>=7.0.0